Move st.image image tests out of "streamlit_test" and into "image_test"
